refactor: 도메인을 많이 알고 있는 Header component /common 에서 분리
